% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\title{method discrepancies among coders}
\usage{
method(path, start, end)
}
\arguments{
\item{path}{the folder where each pubmed .csv export files are stored}

\item{start}{the link.id to begin checking, put in quotes}

\item{end}{the link.id to end checking, put in quotes}
}
\description{
merge together coder sheets\cr
identify and only keep discrepant rows \cr
}
\examples{
method(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep", start = "a21_1", end = "a30b_1")

}
