% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_stats.R
\name{method_stats}
\alias{method_stats}
\title{method discrepancies aggregates}
\usage{
method_stats(path, start, end)
}
\arguments{
\item{path}{the folder where the raw jp method file is and the edited discrepancy file}

\item{start}{the link.id to begin checking, put in quotes}

\item{end}{the link.id to end checking, put in quotes}
}
\description{
method discrepancies aggregates
}
\examples{
method(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep", start = "a21_1", end = "a30b_1")

}
