% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlescreen.R
\name{titlescreen1}
\alias{titlescreen1}
\title{title screening specific for round 1 on sleep and inflammation}
\usage{
titlescreen1(
  path,
  jc.filename,
  jc.sheetnumber,
  pattern,
  recode = F,
  recode.ra = NULL
)
}
\description{
tasks performed:\cr
1.	merge together jc and ra responses\cr
2.	identify and only keep rows with discrepant codes \cr
3.	arrange the discrepant codes by jc 0 vs. ra 1/2; jc 1 vs. ra 0/2; jc 2 vs. ra 0/1\cr
4.	grab ra’s name from filename and place as column name\cr
5.	compute summary descriptives:\itemize{
\item a.	total number of 0, 1, 2 by coder
\item b.	total number of discrepancies
\item c.	total number of each combination of discrepancies
}
6.	repeat above steps for each identified ra file\cr
7.	for recode runs, the discrepancy summary and excel file will be based on second-round values only (i.e., only checking discrepancy in the recode file)\cr
}
\examples{
\strong{no recode has happened yet (use original jc coding):}
titlescreen1(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep/reliability",
jc.filename = "0_merged els & sleep_11.1.19_jc.xlsx",
jc.sheetnumber = 2,
pattern="1_merged els & sleep_11.1.19_ra")

\strong{recode has happened (use jc second round coding):}
titlescreen1(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep/reliability",
jc.filename = "0_merged els & sleep_11.1.19_jc.xlsx",
jc.sheetnumber = 2,
pattern="1_merged els & sleep_11.1.19_ra",
recode=T,
recode.ra = "pk")


}
