% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeclean.R
\name{mergeclean}
\alias{mergeclean}
\title{merge and remove duplicate from pubmed searches}
\usage{
mergeclean(path, newfilename)
}
\arguments{
\item{path}{the folder where each pubmed .csv export files are stored}

\item{newfilename}{the name of the two new consolidated excel files without the extension}
}
\description{
merge together pubmed searches, add columns that index search phase, search terms, and search date\cr
flag the duplicates, export an excel file\cr
remove the duplicates, add a searchid column, export another excel file\cr
video tutorial:\href{https://youtu.be/3zj1b8jar28}{https://youtu.be/3zj1b8jar28}
}
\examples{
mergeclean(path="/Users/phoebelam/Google Drive/2021/4_meta_els sleep & inflammation/1_search_els & sleep",
newfilename = "0_merged els & sleep")

}
