% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\title{method discrepancies among coders}
\usage{
method(path)
}
\arguments{
\item{path}{the folder where each pubmed .csv export files are stored}

\item{newfilename}{the name of the two new consolidated excel files without the extension}
}
\description{
merge together coder sheets\cr
identify and only keep discrepant rows \cr
compute summary descriptives of # discrepancies
}
\examples{
mergeclean(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep",
newfilename = "0_merged els & sleep")

}
