% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlescreen.R
\name{titlescreen1}
\alias{titlescreen1}
\title{title screening specific for round 1 on sleep and inflammation}
\usage{
titlescreen1(
  path,
  jc.filename,
  jc.sheetnumber,
  pattern,
  recode = F,
  recode.ra = NULL
)
}
\arguments{
\item{path}{the folder where both jc and ra files are located (ra files can be in subfolder of this folder)}

\item{jc.filename}{the filename of the jc file including the extension}

\item{jc.sheetnumber}{without quotations, the sheet number counted from the left in which jc codes are logged. include hidden sheets in your count.}

\item{pattern}{pattern by which the function will look for ra files; ensure that the pattern exclusively applies to ra files, excluding the template file}

\item{recode}{FALSE or F: will use jc codes from jc.filename specified above. FALSE is the default value when unspecified; TRUE or T: will use jc codes from the recode file. function will look for the pattern “_recode”, so make sure those exact characters are in filename and make sure this recode file is inside the folder specified as path earlier}

\item{recode.ra}{the initial of the ra (must match the ones on their filename). NULL is the default value when unspecified.}
}
\description{
tasks performed:\cr
1.	merge together jc and ra responses\cr
2.	identify and only keep rows with discrepant codes \cr
3.	arrange the discrepant codes by jc 0 vs. ra 1/2; jc 1 vs. ra 0/2; jc 2 vs. ra 0/1\cr
4.	grab ra’s name from filename and place as column name\cr
5.	compute summary descriptives:\itemize{
\item a.	total number of 0, 1, 2 by coder
\item b.	total number of discrepancies
\item c.	total number of each combination of discrepancies
}
6.	repeat above steps for each identified ra file\cr
7.	for recode runs, the discrepancy summary and excel file will be based on second-round values only (i.e., only checking discrepancy in the recode file)\cr
\cr
video tutorial: \href{https://youtu.be/aQQVP_KCJck}{https://youtu.be/aQQVP_KCJck}
}
\examples{
\strong{no recode has happened yet (use original jc coding):}
titlescreen1(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep/reliability",
jc.filename = "0_merged els & sleep_11.1.19_jc.xlsx",
jc.sheetnumber = 2,
pattern="1_merged els & sleep_11.1.19_ra")

\strong{recode has happened (use jc second round coding):}
titlescreen1(path="/Users/phoebelam/Google Drive/2020/4_meta_els sleep & inflammation/1_search_els & sleep/reliability",
jc.filename = "0_merged els & sleep_11.1.19_jc.xlsx",
jc.sheetnumber = 2,
pattern="1_merged els & sleep_11.1.19_ra",
recode=T,
recode.ra = "pk")


}
